venueiq-backend/requirements.txt
google-adk>=1.18.0
google-cloud-aiplatform>=1.70.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
firebase-admin>=6.5.0
pydantic>=2.0.0
twilio>=9.0.0
python-dotenv>=1.0.0
qrcode[pil]>=7.4.0
httpx>=0.27.0
